NAVADA-Trading/requirements.txt
alpaca-py>=0.28.0
pandas>=2.0.0
numpy>=1.24.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
